+++ WHATS_NEW.md
@@ -1,7 +1,8 @@
 ### 1.0.4
-+ [x] Delegators info
++ [x] Delegations info
 + [x] Mina price info
 + [x] Calculate balance in USD
++ [x] Check for forward fork
 
 ### 1.0.3
 + [x] Five types of alert: `Long-non-sync`, `Diff-with-max`, `Diff-with-unvalidated`, `Hanging-state`, `Diff-with-Explorer`